relocate interface classes
